Moves the arity check to outer scope of c2k func
Moves the check to only occur once on setup instead of on every request.
I wasn't able to figure out how to do this with an optional callback
parameter or the `hasNoCallback` predicate function that's removed here,
as the type checker wasn't able to narrow the type correctly. I worked
around it by having a noop function and making the callback parameter
required.

Also removes the "exclude" from the base tsconfig so that the type
checker works correctly with IDEs. Added a second tsconfig to control
which files emit .js and .d.ts files on build.

Fixes the lint-staged glob which wasn't working correctly